requirements.txt
google-adk==1.18.0
mcp_weather_server